Change 49: Extracting SimBudget and SimBoxOpt
SimBudget and SimBoxOpt have been pulled into their own files for the
sake of readability.
Due to circular references, particularly arising from class method
'fromdict', tail imports have been employed as a Python analogue to C++
forward declarations.
